+++ Globals.py
@@ -1,7 +1,7 @@
 import os.path
 
 # Constants
-_VERSION = 'alpha-0.5.1'
+_VERSION = 'alpha-0.5.2'
 _TITLE = f"ToNCodes {_VERSION} - by MichelleDB"
 _GITHUB = 'https://github.com/69MichelleDB/ToNCodes'
 
@@ -38,6 +38,7 @@ newCodeAdded = True
 killersList = []
 
 # Round variables
+roundEvent = ''
 roundMap = ''
 roundKiller = ''
 roundCondition = ''